Fix bug in to-message conversion handler for Transform, Pose.
This commit fixes a bug in the to-message conversion handler for
the message types `Transform` and `Pose` from the geometry_msgs
package. Previous to this commit, attempting to convert a
NumPy representation into one of these message types would have
yielded a `TypeError` due to an incorrect method call. Also,
the handler has been erroneously registered for the types
`TransformStamped`, `PoseStamped`.